# examples/variants/json/target/openapi.json
{
  "components": {
    "parameters": {
      "Limit": {
        "description": "Maximum number of items to return.",
        "in": "query",
        "name": "limit",
        "required": false,
        "schema": {
          "default": 20,
          "format": "int32",
          "maximum": 100,
          "minimum": 1,
          "type": "integer"
        }
      },
      "Offset": {
        "description": "Number of items to skip.",
        "in": "query",
        "name": "offset",
        "required": false,
        "schema": {
          "default": 0,
          "format": "int32",
          "minimum": 0,
          "type": "integer"
        }
      },
      "ProjectId": {
        "description": "Project identifier.",
        "in": "path",
        "name": "project_id",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      },
      "TaskId": {
        "description": "Task identifier.",
        "in": "path",
        "name": "task_id",
        "required": true,
        "schema": {
          "format": "uuid",
          "type": "string"
        }
      }
    },
    "requestBodies": {
      "CommentCreate": {
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "attachmentUrls": {
                  "items": {
                    "format": "uri",
                    "type": "string"
                  },
                  "type": "array"
                },
                "body": {
                  "maxLength": 2000,
                  "minLength": 1,
                  "type": "string"
                }
              },
              "required": [
                "body"
              ],
              "type": "object"
            }
          }
        },
        "required": true
      }
    },
    "responses": {
      "ActivityEventListRespObj": {
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "items": {
                  "items": {
                    "$ref": "#/components/schemas/ActivityEventObj"
                  },
                  "type": "array"
                },
                "nextCursor": {
                  "nullable": true,
                  "type": "string"
                }
              },
              "required": [
                "items",
                "nextCursor"
              ],
              "type": "object"
            }
          }
        },
        "description": "Activity events."
      },
      "CommentListRespObj": {
        "content": {
          "application/json": {
            "schema": {
              "items": {
                "$ref": "#/components/schemas/CommentObj"
              },
              "type": "array"
            }
          }
        },
        "description": "Comments for a task."
      },
      "ProblemRespObj": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DefaultErrorObj"
            }
          }
        },
        "description": "Error response."
      }
    },
    "schemas": {
      "ActivityEventObj": {
        "properties": {
          "action": {
            "enum": [
              "created",
              "updated",
              "deleted",
              "commented"
            ],
            "type": "string"
          },
          "actor": {
            "$ref": "#/components/schemas/UserSummaryObj"
          },
          "changes": {
            "additionalProperties": {
              "properties": {
                "after": {},
                "before": {}
              },
              "type": "object"
            },
            "type": "object"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "target": {
            "properties": {
              "id": {
                "format": "uuid",
                "type": "string"
              },
              "type": {
                "enum": [
                  "task",
                  "project",
                  "comment"
                ],
                "type": "string"
              }
            },
            "required": [
              "type",
              "id"
            ],
            "type": "object"
          }
        },
        "required": [
          "id",
          "actor",
          "action",
          "target",
          "createdAt"
        ],
        "type": "object"
      },
      "CommentObj": {
        "properties": {
          "attachments": {
            "items": {
              "properties": {
                "metadata": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "name": {
                  "type": "string"
                },
                "url": {
                  "format": "uri",
                  "type": "string"
                }
              },
              "required": [
                "name",
                "url"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "author": {
            "$ref": "#/components/schemas/UserSummaryObj"
          },
          "body": {
            "maxLength": 2000,
            "minLength": 1,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "author",
          "body",
          "createdAt"
        ],
        "type": "object"
      },
      "DefaultErrorObj": {
        "properties": {
          "details": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "error": {
            "description": "Machine-readable error code.",
            "example": "invalid_request",
            "type": "string"
          },
          "message": {
            "description": "Human-readable message.",
            "type": "string"
          },
          "requestId": {
            "description": "Request correlation id.",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "type": "object"
      },
      "ErrorObj": {
        "properties": {
          "code": {
            "description": "Machine-readable error code.",
            "format": "int32",
            "type": "integer"
          },
          "details": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "message": {
            "description": "Human-readable message.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "LabelObj": {
        "properties": {
          "color": {
            "description": "Hex color, e.g. #1a2b3c.",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "type": "string"
          },
          "name": {
            "maxLength": 40,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "LoginRequestObj": {
        "properties": {
          "email": {
            "format": "email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "minLength": 8,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
      "LoginResponseObj": {
        "properties": {
          "expiresIn": {
            "description": "Token lifetime in seconds.",
            "format": "int32",
            "type": "integer"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "expiresIn"
        ],
        "type": "object"
      },
      "NewTaskObj": {
        "properties": {
          "labels": {
            "items": {
              "$ref": "#/components/schemas/LabelObj"
            },
            "type": "array"
          },
          "projectId": {
            "format": "uuid",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "title",
          "projectId"
        ],
        "type": "object"
      },
      "ProjectObj": {
        "properties": {
          "defaultLabel": {
            "$ref": "#/components/schemas/LabelObj"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "type": "object"
      },
      "StatusObj": {
        "properties": {
          "status": {
            "enum": [
              "ok",
              "degraded"
            ],
            "type": "string"
          },
          "uptimeSeconds": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
      "TaskObj": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "labels": {
            "items": {
              "$ref": "#/components/schemas/LabelObj"
            },
            "type": "array"
          },
          "project": {
            "$ref": "#/components/schemas/ProjectObj"
          },
          "status": {
            "enum": [
              "open",
              "in_progress",
              "done"
            ],
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "project"
        ],
        "type": "object"
      },
      "UserSummaryObj": {
        "properties": {
          "avatarUrl": {
            "format": "uri",
            "type": "string"
          },
          "displayName": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "format": "email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKeyHeader": {
        "description": "Static API key for service-to-service calls.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "Bearer JWT in the Authorization header.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Small hand-written API used as the goopenapigen example. paths/ and components/ are merged\nby folder convention, so the root stays a thin envelope and nothing is wired by hand.\n",
    "title": "Tasks API",
    "version": "1.3.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/auth/login": {
      "post": {
        "operationId": "authLogin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestObj"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseObj"
                }
              }
            },
            "description": "Issued token."
          },
          "401": {
            "$ref": "#/components/responses/ProblemRespObj"
          }
        },
        "security": [],
        "summary": "Exchange credentials for a token.",
        "tags": [
          "auth"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "healthGet",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusObj"
                }
              }
            },
            "description": "Service status."
          }
        },
        "security": [],
        "summary": "Liveness probe.",
        "tags": [
          "system"
        ]
      }
    },
    "/projects/{project_id}/activity": {
      "get": {
        "operationId": "projectActivityList",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ActivityEventListRespObj"
          },
          "default": {
            "$ref": "#/components/responses/ProblemRespObj"
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          }
        ],
        "summary": "List project activity events.",
        "tags": [
          "projects",
          "activity"
        ]
      }
    },
    "/tasks": {
      "get": {
        "operationId": "tasksList",
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TaskObj"
                  },
                  "type": "array"
                }
              }
            },
            "description": "A page of tasks."
          },
          "default": {
            "$ref": "#/components/responses/ProblemRespObj"
          }
        },
        "summary": "List tasks.",
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "tasksCreate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewTaskObj"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskObj"
                }
              }
            },
            "description": "Created task."
          },
          "default": {
            "$ref": "#/components/responses/ProblemRespObj"
          }
        },
        "summary": "Create a task.",
        "tags": [
          "tasks"
        ]
      }
    },
    "/tasks/{task_id}": {
      "delete": {
        "operationId": "taskDelete",
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "default": {
            "$ref": "#/components/responses/ProblemRespObj"
          }
        },
        "summary": "Delete a task.",
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "operationId": "taskGet",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskObj"
                }
              }
            },
            "description": "The task."
          },
          "default": {
            "$ref": "#/components/responses/ProblemRespObj"
          }
        },
        "summary": "Get a task by id.",
        "tags": [
          "tasks"
        ]
      },
      "parameters": [
        {
          "$ref": "#/components/parameters/TaskId"
        }
      ]
    },
    "/tasks/{task_id}/comments": {
      "get": {
        "operationId": "taskCommentsList",
        "parameters": [
          {
            "$ref": "#/components/parameters/TaskId"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/CommentListRespObj"
          },
          "default": {
            "$ref": "#/components/responses/ProblemRespObj"
          }
        },
        "summary": "List task comments.",
        "tags": [
          "tasks",
          "comments"
        ]
      },
      "post": {
        "operationId": "taskCommentsCreate",
        "parameters": [
          {
            "$ref": "#/components/parameters/TaskId"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/CommentCreate"
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentObj"
                }
              }
            },
            "description": "Created comment."
          },
          "default": {
            "$ref": "#/components/responses/ProblemRespObj"
          }
        },
        "summary": "Add a comment to a task.",
        "tags": [
          "tasks",
          "comments"
        ]
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://{region}.api.example.com/{basePath}",
      "variables": {
        "basePath": {
          "default": "v1",
          "description": "API base path."
        },
        "region": {
          "default": "eu",
          "description": "Deployment region.",
          "enum": [
            "eu",
            "us"
          ]
        }
      }
    }
  ],
  "tags": [
    {
      "description": "Task management.",
      "name": "tasks"
    },
    {
      "description": "Authentication.",
      "name": "auth"
    },
    {
      "description": "Service health.",
      "name": "system"
    },
    {
      "description": "Project-scoped operations.",
      "name": "projects"
    },
    {
      "description": "Task discussions.",
      "name": "comments"
    },
    {
      "description": "Audit and activity streams.",
      "name": "activity"
    }
  ]
}